requirements.txt
flask
flask-sqlalchemy
sqlalchemy-migrate
flask-wtf
flask-admin
flask-bootstrap
flask-nav